LibWeb: Split OverflowStyleValue out of StyleValue.{h,cpp}
