+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -89,6 +89,7 @@ set(SOURCES
     CSS/StyleValues/LinearGradientStyleValue.cpp
     CSS/StyleValues/ListStyleStyleValue.cpp
     CSS/StyleValues/NumericStyleValue.cpp
+    CSS/StyleValues/OverflowStyleValue.cpp
     CSS/StyleValues/RadialGradientStyleValue.cpp
     CSS/Supports.cpp
     CSS/SyntaxHighlighter/SyntaxHighlighter.cpp
